refactor(framework): one node:fs adapter behind the four factories
nodeStoreFs, nodeRegistryFs, nodeMaintenanceFs and nodeProjectFs each
hand-rolled the same node:fs/promises adapter, dynamic-importing the
module per call. read/write/mkdir were byte-identical in the three that
have them, and the comments already admitted it: each said "same
convention as" and pointed at a different sibling, so the convention was
documented in a circle rather than shared.

One nodeFs() in node-fs.ts holds every primitive; each factory returns
the slice its own interface declares. The four narrow interfaces stay:
they are the consumers' contracts and deliberately small (a store may
append, a registry may not). Only the implementation is shared. 13
dynamic imports across four files become 7 in one.

Each factory destructures rather than returning nodeFs() whole, so the
object it hands back carries exactly the methods it did before. Returning
it whole passed every test but widened the runtime surface, which the
probe caught and the suite did not.

Proven by exercising all 7 primitives against a real temp dir, before and
after: identical, including read(missing) rejecting ENOENT, exists(dir)
being false, readdir(missing) yielding [], and isDirectory being
stat().isDirectory() rather than exists's stat().isFile().

The client barrel still reaches no node: import (#431/#520 guard passes);
none of the four is reachable from it.

Closes #583
